+++ src/components/ProductCardList/product.type.ts
