+++ src/components/ShoppingCartModal/cart.type.ts
@@ -1,4 +1,4 @@
-import { Product } from './product.type';
+import { Product } from '../ProductCardList/product.type';
 
 interface CartItem {
   id: number;
