+++ src/contexts/context.constant.ts
@@ -1,4 +1,4 @@
-import { ErrorState } from '../types/error.type';
+import { ErrorState } from '../components/ErrorToast/error.type';
 
 export const INITIAL_ERROR: ErrorState = {
   is: false,
